benchmarks/personnel_cards/correction_ui/requirements_flask.txt
Flask>=3.0.0
Pillow>=10.0.0
